backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai==0.8.3
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.2.2
pandas==2.2.3
aiofiles==24.1.0
python-dotenv==1.0.1
